Move modules in the main package to cmd package
* Moved the content in the main package to cmd package. The erstwhile
  main.go becomes legacy_main.go.
* This change is aimed to ease the move to Cobra + Viper for config
  management.
* A Cobra app has usually a very lightweight main function and delegates
  to cmd for the logic. We plan to follow the same template.
* This refactoring will help share common logic between the
  legacy_main.go and root.go.
